Note certificate
They say in their docs to assume the certificate is authenticated since they say the CA is out of the scope of this project. However, we disagree and think some server authentication is needed. But, we can still break it even if there were a CA since we can forge digital signatures.